+++ decorator.rb
@@ -10,19 +10,3 @@ class Decorator < Nameable
     @nameable.correct_name
   end
 end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    if @nameable.correct_name.length > 10
-      @nameable.correct_name[0..9]
-    else
-      @nameable.correct_name
-    end
-  end
-end
